// src/app/types/attachmentsType.ts
export type AttachmentType = {
    attachmentType: string
    base64: string
    contractId: number
    createBy?: number | null
    createDate: string
    fileId: string
    fileName: string
    id: number
    mimeType: string
    modBy?: number | null
    modDate: string
    size: number
    type?: number | null
}